Moved UIkit type definitions
